Reorganise python code
I don't know whether this is the correct way to structure a python
project, but it was getting a bit messy in that dir so I have moved
things around. Also the Welder package can probably be refactored one day.
